% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ps_installed.R
\name{verify_ps_installed}
\alias{verify_ps_installed}
\title{Check if ps-eval is available}
\usage{
verify_ps_installed()
}
\value{
\code{ìnvisible(NULL)} if \code{ps-eval} is available. Otherwise, an error
is thrown.
}
\description{
This function throws an error if \code{system("ps-eval")} is not available.
The error message indicates to the user where installation instructions
for pokerstove can be found.
}
